Refactor des hydrators Ting
- déplacement vers notre namespace Ting
- ajout de types natifs
